Update get_app_path() to handle /proc weirdness.
On Linux, (and thus, Android) attempting to lstat() a symlink under
/proc succeeds, but returns a .st_size value of 0, making it
unsuitable for use in attempting to determine the buffer size of
the actual path string. As a workaround, we currently write the
first PATH_MAX characters from the path into the buffer, and then
append a NUL character at the end. This obviously isn't sufficient
for really long paths, but it's good enough for what is needed at
the moment.